Update SDCard.cpp
Adding a line to specify the SDSPI max frequency. Came across an issue where my Lolin32 esp32 couldn't detect the SD card on the default frequency so had to specify to a lower one to make it work. The choice of frequency came from seeing what is used in the SDFat Library for ESP's. 

I suspect the issue is more to do with the breakout board used for the micro SD than the esp32 or SD cards themselves. Limiting it to the lower speed should make it compatible with most (hopefully all) boards and can be increased if desired.